Fix all lint errors returned by golangci-lint
We don't care that the md4 algorithm isn't cryptographically secure, we
are not using it for anything like that.

I've handled all appropriate errors from io.Copy and io.CopyN and
explicitly ignored when not appropriate.

I've also renamed the testing.T and testing.B variables in the tests to
be what they are usually named.